+++ backend/api/models/utils_models.py
@@ -299,13 +299,13 @@ class MediaLabelMixin:
 
     @classmethod
     def get_user_labels(cls, user_id: int) -> List[str]:
-        q_all = db.session.query(cls.label.distinct()).filter_by(user_id=user_id).all()
+        q_all = db.session.query(cls.label.distinct()).filter_by(user_id=user_id).order_by(cls.label).all()
         return [label[0] for label in q_all]
 
     @classmethod
     def get_user_media_labels(cls, user_id: int, media_id: int) -> Dict:
         all_labels = set(cls.get_user_labels(user_id))
-        q_in = db.session.query(cls.label).filter_by(user_id=user_id, media_id=media_id).all()
+        q_in = db.session.query(cls.label).filter_by(user_id=user_id, media_id=media_id).order_by(cls.label).all()
         already_in = {label[0] for label in q_in}
         available = all_labels - already_in
         return dict(already_in=list(already_in), available=list(available))