+++ frontend/src/main.jsx
@@ -1,8 +1,8 @@
 import {routeTree} from "./routeTree.gen";
 import {createRoot} from "react-dom/client";
-import {ErrorPage} from "@/pages/ErrorPage";
 import {initializeUserClient} from "@/api/MyApiClient";
 import {ThemeProvider} from "@/providers/ThemeProvider";
+import {ErrorComponent} from "@/components/app/base/ErrorComponent";
 import {createRouter, RouterProvider} from "@tanstack/react-router";
 import "./index.css";
 
@@ -11,7 +11,7 @@ const router = createRouter({
     defaultGcTime: 0,
     routeTree: routeTree,
     defaultPreload: false,
-    defaultNotFoundComponent: ErrorPage,
+    defaultNotFoundComponent: ErrorComponent,
     defaultErrorComponent: DefaultErrorComponent,
 });
 
@@ -41,9 +41,9 @@ function App() {
 
 function DefaultErrorComponent({ error }) {
     try {
-        return <ErrorPage {...JSON.parse(error.message)}/>;
+        return <ErrorComponent {...JSON.parse(error.message)}/>;
     }
     catch {
-        return <ErrorPage/>;
+        return <ErrorComponent/>;
     }
 }